supabase_client.py

from supabase import create_client
import os
from dotenv import load_dotenv

load_dotenv()
url = os.getenv("SUPABASE_URL")
key = os.getenv("SUPABASE_KEY")
supabase = create_client(url, key)

def insert_to_supabase(table, data):
    return supabase.table(table).insert(data).execute()
