requirements.txt
pdfplumber
openai
python-dotenv
supabase
